feat: prompts for onRemove

+++ src/components/Aside/DocList.js
@@ -6,28 +6,28 @@ export default function DocList({
   $target,
   initialState = {
     documents: [],
-    openedDocId
+    openedDocId,
   },
   onSelect,
   onAdd,
   onRemove,
 }) {
-  const $listContainer = makeElement('nav', 'doc-list');
+  const $listContainer = makeElement("nav", "doc-list");
   $target.appendChild($listContainer);
 
   const $list = makeUl("root");
-  const LIST_SCROLLTOP_SAVE_KEY = 'listScrollPos'
+  const LIST_SCROLLTOP_SAVE_KEY = "listScrollPos";
 
   this.state = initialState;
 
   this.setState = (nextState) => {
     this.state = nextState;
-    console.log(this.state)
+    console.log(this.state);
     this.render();
   };
 
   this.render = () => {
-    $list.innerHTML = '';
+    $list.innerHTML = "";
     createList($list, this.state.documents);
     $listContainer.appendChild($list);
     goToListScrollPos();
@@ -41,8 +41,8 @@ export default function DocList({
         const { documents, id } = value;
         const $li = makeLi(value);
 
-        if(this.state.openedDocId === id.toString()) {
-          addClass($li, 'on');
+        if (this.state.openedDocId === id.toString()) {
+          addClass($li, "on");
         }
 
         const hasChild = documents.length > 0;
@@ -51,10 +51,10 @@ export default function DocList({
           $ul.dataset.parentId = id;
           createList($ul, documents);
 
-          if($ul.querySelector('li.on')) {
-            addClass($ul, 'on');
+          if ($ul.querySelector("li.on")) {
+            addClass($ul, "on");
           }
-          
+
           $li.appendChild($ul);
         }
         $target.appendChild($li);
@@ -64,16 +64,16 @@ export default function DocList({
 
   const goToListScrollPos = () => {
     var listScrollPos = session.getItem(LIST_SCROLLTOP_SAVE_KEY, 0);
-    if(listScrollPos > 0) {
+    if (listScrollPos > 0) {
       console.log(listScrollPos);
       $list.scrollTo(0, listScrollPos);
       session.removeItem(LIST_SCROLLTOP_SAVE_KEY);
     }
-  }
+  };
 
-  window.addEventListener('beforeunload', () => {
-    session.setItem(LIST_SCROLLTOP_SAVE_KEY, $list.scrollTop)
-  })
+  window.addEventListener("beforeunload", () => {
+    session.setItem(LIST_SCROLLTOP_SAVE_KEY, $list.scrollTop);
+  });
 
   $listContainer.addEventListener("click", (e) => {
     e.preventDefault();
@@ -84,32 +84,40 @@ export default function DocList({
 
     const $target = e.target;
     const { tagName, className } = $target;
-    const splitedClassName = className.split(' ').filter(x => !x.startsWith('xi')).shift();
+    const splitedClassName = className
+      .split(" ")
+      .filter((x) => !x.startsWith("xi"))
+      .shift();
 
-    if(tagName === 'I') {
+    if (tagName === "I") {
       switch (splitedClassName) {
-        case 'view-more':
-          const $parent = $li.querySelector('.parent');
-          if(!$parent) return;
+        case "view-more":
+          const $parent = $li.querySelector(".parent");
+          if (!$parent) return;
 
-          if(hasClass($parent, 'on')) {
-            removeClass($target, 'on');
-            removeClass($parent, 'on');
+          if (hasClass($parent, "on")) {
+            removeClass($target, "on");
+            removeClass($parent, "on");
           } else {
-            addClass($target, 'on');
-            addClass($parent, 'on');
+            addClass($target, "on");
+            addClass($parent, "on");
           }
           break;
-        case 'add':
+        case "add":
           onAdd(documentId);
           break;
-        case 'remove':
-          if(hasClass($li, 'on')) return;
-          onRemove(documentId);
+        case "remove":
+          if (hasClass($li, "on")) {
+            alert('Currently open! Cannot delete 🙅‍♀️');
+            return;
+          }
+          if(confirm('Are you sure you want to delete? 🚮')) onRemove(documentId);
           break;
       }
-    } else if (tagName === 'A' && splitedClassName === 'title') {
-      document.querySelectorAll('.list-item').forEach($li => $li.classList.remove('on'));
+    } else if (tagName === "A" && splitedClassName === "title") {
+      document
+        .querySelectorAll(".list-item")
+        .forEach(($li) => $li.classList.remove("on"));
       onSelect(documentId);
     }
   });
